make it IDE friendly
- extract shared variables and functions so they will be visible from production code
- use the test framework to fail so that the IDE understands that the test fails and ends
