+++ go.mod
@@ -1,0 +1,3 @@
+module go_koans
+
+go 1.17
